edge-medical-agent/requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LangGraph
langgraph>=0.0.40
langchain>=0.1.0

# Async HTTP
httpx>=0.26.0
aiohttp>=3.9.0

# Database clients
motor>=3.3.0  # Async MongoDB
redis>=5.0.0  # Redis client
opensearch-py>=2.4.0  # OpenSearch logging

# Scheduling
apscheduler>=3.10.0

# PDF processing
pypdf>=3.17.0
pdf2image>=1.16.0
Pillow>=10.0.0

# OCR (for scanned PDFs)
pytesseract>=0.3.10

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Email
aiosmtplib>=3.0.0

# Data processing
pandas>=2.1.0

# Fuzzy matching (optional, for LOINC fuzzy search)
rapidfuzz>=3.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
hypothesis>=6.92.0  # Property-based testing
httpx>=0.26.0  # For TestClient
